+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 The present file will list all changes made to the project; according to the
 [Keep a Changelog](http://keepachangelog.com/) project.
 
-## [9.4.1] unreleased
+## [9.4.2] unreleased
+
+### API changes
+
+#### Deprecated
+
+The following methods have been deprecated:
+
+- `CommonDBTM::getTablesOf()`
+- `CommonDBTM::getForeignKeyFieldsOf()`
+
+## [9.4.1] 2019-03-15
 
 ### API changes